sf-express/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
cryptography==46.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
requests==2.31.0
httpx==0.27.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
hypothesis==6.92.1